refactor(User): fixing up user utils and credential typing.
