+++ src/modules/CompletedChallenges/Widget/CompletedChallengesWidget.selector.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { applySpec, map, pick, pipe, slice } from 'ramda'
 
-import * as UserConstants from '~/modules/User/User.constants'
+import { USER_OPPORTUNITY_CREDENTIAL_WIDGET_SELECTOR_SPEC } from '~/modules/User/User.constants'
 import { selectUserChallenges } from '~/modules/UserChallenges/UserChallenges.selector'
 
 export default createSelector(selectUserChallenges, userChallenges => {
@@ -9,7 +9,7 @@ export default createSelector(selectUserChallenges, userChallenges => {
   const ids = slice(0, 2, userChallenges.ids)
   const entities = pipe(
     pick(ids),
-    map(applySpec(UserConstants.USER_OPPORTUNITY_CREDENTIAL_WIDGET_SELECTOR_SPEC)),
+    map(applySpec(USER_OPPORTUNITY_CREDENTIAL_WIDGET_SELECTOR_SPEC)),
   )(userChallenges.entities)
 
   return { userChallenges: { ids, entities }, count }
