+++ src/modules/UserEducation/UserEducation.middleware.ts
@@ -6,7 +6,7 @@ import { Middleware } from 'redux'
 import { actions as ApiActions, utils as ApiUtils } from '~/api'
 import { constants as ApiUsersConstants, types as ApiUsersTypes } from '~/api/users'
 import * as Strings from '~/constants/strings.constants'
-import { actions as EducationActions } from '~/modules/Education'
+import * as EducationActions from '~/modules/Education/Education.reducer'
 import { utils as ErrorUtils } from '~/modules/Error'
 import { HomeNavigationRoutes } from '~/modules/HomeNavigation/HomeNavigation.types'
 import { types as NavigationTypes, utils as NavigationUtils } from '~/modules/Navigation'
